refactor: cli delegates sc calls to api
