+++ api/bin/chainflip-cli/Cargo.toml
@@ -7,6 +7,7 @@ version = "0.5.0"
 [dependencies]
 anyhow = "1.0"
 chainflip-engine = {path = "../../../engine/"}
+chainflip-api = {path = "../../lib"}
 clap = {version = "3.1.18", features = ["derive"]}
 config = "0.13.1"
 futures = "0.3.14"
@@ -15,24 +16,9 @@ serde = {version = "1.0", features = ["derive", "rc"]}
 slog = {version = "2.7.0", features = ["max_level_trace", "release_max_level_trace"]}
 tokio = {version = "1.13.1", features = ["full"]}
 utilities = {path = "../../../utilities"}
-web3 = {git = 'https://github.com/tomusdrw/rust-web3.git', rev = '8796c88'}
-jsonrpsee = {version = "0.13.0", features = ["client"]}
-async-trait = "0.1.49"
 
 # State chain
-cf-chains = {path = "../../../state-chain/chains"}
 chainflip-node = {path = "../../../state-chain/node"}
-pallet-cf-environment = {path = "../../../state-chain/pallets/cf-environment"}
-pallet-cf-staking = {path = "../../../state-chain/pallets/cf-staking"}
-pallet-cf-threshold-signature = {path = "../../../state-chain/pallets/cf-threshold-signature"}
-pallet-cf-validator = {path = "../../../state-chain/pallets/cf-validator"}
-pallet-cf-account-roles = { path = '../../../state-chain/pallets/cf-account-roles' }
 cf-primitives = {path = "../../../state-chain/primitives"}
 pallet-cf-governance = {path = "../../../state-chain/pallets/cf-governance"}
 state-chain-runtime = {path = "../../../state-chain/runtime"}
-frame-system = {git = 'https://github.com/chainflip-io/substrate.git', tag = 'chainflip-monthly-2022-06+01'}
-
-# Substrate key types
-sp-consensus-aura = {git = 'https://github.com/chainflip-io/substrate.git', tag = 'chainflip-monthly-2022-06+01'}
-sp-core = {git = 'https://github.com/chainflip-io/substrate.git', tag = 'chainflip-monthly-2022-06+01'}
-sp-finality-grandpa = {git = 'https://github.com/chainflip-io/substrate.git', tag = 'chainflip-monthly-2022-06+01'}
