+++ api/lib/Cargo.toml
@@ -6,3 +6,27 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+anyhow = "1.0"
+hex = "0.4.3"
+slog = {version = "2.7.0", features = ["max_level_trace", "release_max_level_trace"]}
+web3 = {git = 'https://github.com/tomusdrw/rust-web3.git', rev = '8796c88'}
+futures = "0.3.14"
+async-trait = "0.1.49"
+chainflip-engine = {path = "../../engine/"}
+
+# State Chain
+state-chain-runtime = {path = "../../state-chain/runtime"}
+cf-chains = {path = "../../state-chain/chains"}
+pallet-cf-environment = {path = "../../state-chain/pallets/cf-environment"}
+cf-primitives = {path = "../../state-chain/primitives"}
+pallet-cf-staking = {path = "../../state-chain/pallets/cf-staking"}
+pallet-cf-validator = {path = "../../state-chain/pallets/cf-validator"}
+pallet-cf-governance = {path = "../../state-chain/pallets/cf-governance"}
+pallet-cf-threshold-signature = {path = "../../state-chain/pallets/cf-threshold-signature"}
+pallet-cf-account-roles = { path = '../../state-chain/pallets/cf-account-roles' }
+frame-system = {git = 'https://github.com/chainflip-io/substrate.git', tag = 'chainflip-monthly-2022-06+01'}
+
+# Substrate key types
+sp-consensus-aura = {git = 'https://github.com/chainflip-io/substrate.git', tag = 'chainflip-monthly-2022-06+01'}
+sp-core = {git = 'https://github.com/chainflip-io/substrate.git', tag = 'chainflip-monthly-2022-06+01'}
+sp-finality-grandpa = {git = 'https://github.com/chainflip-io/substrate.git', tag = 'chainflip-monthly-2022-06+01'}
